+++ hslua-objectorientation/src/HsLua/ObjectOrientation/ListType.hs
@@ -1,17 +1,23 @@
+{-# LANGUAGE FlexibleInstances        #-}
+{-# LANGUAGE MultiParamTypeClasses    #-}
 {-# LANGUAGE OverloadedStrings        #-}
 {-# LANGUAGE ScopedTypeVariables      #-}
 {-# LANGUAGE TypeApplications         #-}
 module HsLua.ObjectOrientation.ListType
-  ( ListSpec
-    -- * Semi-internal
-  , lazyListStateName
-  , lazylisteval
-  , setList
+  ( UDTypeWithList
+  , ListSpec
+  , listExtension
   ) where
 
 import Control.Monad ((<$!>), forM_, void)
-import HsLua.Core
+import HsLua.Core as Lua
 import HsLua.Marshalling
+import HsLua.ObjectOrientation.Generic
+
+-- | Userdata type that (also) behaves like a list.
+type UDTypeWithList e fn a itemtype =
+  UDTypeGeneric e fn a
+{-# DEPRECATED UDTypeWithList "Use UDTypeGeneric instead" #-}
 
 -- | Pair of pairs, describing how a type can be used as a Lua list. The
 -- first pair describes how to push the list items, and how the list is
@@ -22,6 +28,35 @@ type ListSpec e a itemtype =
   ( (Pusher e itemtype, a -> [itemtype])
   , (Peeker e itemtype, a -> [itemtype] -> a)
   )
+
+listExtension
+  :: LuaError e
+  => ListSpec e a itemtype
+  -> UDTypeHooks e fn a
+listExtension ((pushItem, toList), (peekItem, updateList)) =
+  UDTypeHooks
+  { hookMetatableSetup = do
+      pushName "lazylisteval"
+      pushHaskellFunction (lazylisteval pushItem)
+      rawset (nth 3)
+
+  , hookPeekUD = \x idx ->
+    (`lastly` pop 1) $ liftLua (getiuservalue idx 1) >>= \case
+      TypeTable -> setList peekItem updateList x
+      _other    -> pure x
+
+  , hookPushUD = \x -> do
+      newtable
+      pushName "__lazylist"
+      newhsuserdatauv (toList x) 1
+      void (newudmetatable lazyListStateName)
+      setmetatable (nth 2)
+      rawset (nth 3)
+      void (setiuservalue (nth 2) 1)
+
+  , hookUservalues = 1
+  }
+
 
 -- | Evaluate part of a lazy list. Takes the following arguments, in
 -- this order:
@@ -66,10 +101,12 @@ lazyListStateName = "HsLua unevalled lazy list"
 -- | Gets a list from a uservalue table and sets it on the given value.
 -- Expects the uservalue (i.e., caching) table to be at the top of the
 -- stack.
-setList :: forall itemtype e a. LuaError e
-        => ListSpec e a itemtype -> a
+setList :: forall itemtype a e. LuaError e
+        => Peeker e itemtype
+        -> (a -> [itemtype] -> a)
+        -> a
         -> Peek e a
-setList (_pushspec, (peekItem, updateList)) x = (x `updateList`) <$!> do
+setList peekItem updateList x = (x `updateList`) <$!> do
   liftLua (getfield top "__lazylistindex") >>= \case
     TypeBoolean -> do
       -- list had been fully evaluated
@@ -79,7 +116,11 @@ setList (_pushspec, (peekItem, updateList)) x = (x `updateList`) <$!> do
       let getLazyList = do
             liftLua (getfield top "__lazylist") >>= \case
               TypeUserdata -> pure ()
-              _ -> failPeek "unevaled items of lazy list cannot be peeked"
+              otherType -> do
+                tyname <- liftLua $ typename otherType
+                failPeek $
+                  "unevaled items of lazy list cannot be peeked: got " <>
+                  tyname
             (`lastly` pop 1) $ reportValueOnFailure
               lazyListStateName
               (\idx -> fromuserdata @[itemtype] idx lazyListStateName)