+++ hslua-objectorientation/test/HsLua/ObjectOrientationTests.hs
@@ -373,6 +373,7 @@ tests = testGroup "Object Orientation"
     ]
   ]
 
+-- | Define a default UDType without bells and whistles.
 deftype :: LuaError e
         => Name                              -- ^ type name
         -> [(Operation, HaskellFunction e)]  -- ^ operations
@@ -380,16 +381,10 @@ deftype :: LuaError e
         -> UDType e (HaskellFunction e) a
 deftype = deftypeGeneric pushHaskellFunction
 
-deftype' :: LuaError e
-         => Name                  -- ^ type name
-         -> [(Operation, HaskellFunction e)]  -- ^ operations
-         -> [Member e (HaskellFunction e) a]  -- ^ methods
-         -> Maybe (ListSpec e a itemtype)  -- ^ list access
-         -> UDTypeWithList e (HaskellFunction e) a itemtype
-deftype' = deftypeGeneric' pushHaskellFunction
-
 -- | Pushes a userdata value of the given type.
-pushUD :: LuaError e => UDTypeWithList e fn a itemtype -> a -> LuaE e ()
+pushUD
+  :: LuaError e
+  => UDTypeGeneric e fn a -> a -> LuaE e ()
 pushUD = pushUDGeneric (const (pure ()))
 
 -- | Define a (meta) operation on a type.
@@ -432,17 +427,18 @@ newtype LazyIntList = LazyIntList { fromLazyIntList :: [Int] }
   deriving (Eq, Show)
 
 typeLazyIntList :: LuaError e
-                => UDTypeWithList e (HaskellFunction e) LazyIntList Int
-typeLazyIntList = deftype' "LazyIntList"
+                => UDTypeGeneric e (HaskellFunction e) LazyIntList
+typeLazyIntList = deftypeGeneric' pushHaskellFunction "LazyIntList"
   [ operation Tostring $ do
       lazyList <- forcePeek $ peekUDGeneric typeLazyIntList (nthBottom 1)
       pushString (show lazyList)
       return (NumResults 1)
   ]
   [ alias "seq" "sequence" [] ]
-  (Just ( (pushIntegral, fromLazyIntList)
-        , (peekIntegral, \_ lst -> LazyIntList lst)
-        ))
+  (listExtension
+    ( (pushIntegral, fromLazyIntList)
+    , (peekIntegral, \_ lst -> LazyIntList lst)
+    ))
 
 --
 -- Sample sum type